requirements.txt
# Основные зависимости
pyautogui>=0.9.54
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
keyboard>=0.13.5
numpy>=1.24.0

# Конфигурация
PyYAML>=6.0

# Для разработки
pytest>=7.4